pyproject.toml
[tool.poetry]
name = "questionanswer"
version = "0.1.0"
description = ""
authors = ["kirvitchenko <vitchenkokir@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.8"
djangorestframework = "^3.16.1"
black = "^25.11.0"
psycopg2-binary = "^2.9.11"
pytest = "^9.0.1"
pytest-django = "^4.11.1"
pytest-factoryboy = "^2.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
